+++ include/CLI/Option.hpp
@@ -185,9 +185,19 @@ public:
         return this;
     }
 
-    /// Any number supported
-    template<typename... ARG>
-    Option* requires(Option* opt, Option* opt1, ARG... args) {
+    /// Can find a string if needed
+    template<typename T=App>
+    Option* requires(std::string opt_name) {
+        for(const Option_p& opt : dynamic_cast<T*>(parent_)->options_)
+            if(opt.get() != this && opt->check_name(opt_name))
+                return requires(opt.get());
+        throw IncorrectConstruction("Option " + opt_name + " is not defined");
+
+    }
+
+    /// Any number supported, any mix of string and Opt
+    template<typename A, typename B, typename... ARG>
+    Option* requires(A opt, B opt1, ARG... args) {
         requires(opt);
         return requires(opt1, args...);
     }
@@ -200,9 +210,18 @@ public:
         return this;
     }
 
-    /// Any number supported
-    template<typename... ARG>
-    Option* excludes(Option* opt, Option* opt1, ARG... args) {
+    /// Can find a string if needed
+    template<typename T=App>
+    Option* excludes(std::string opt_name) {
+        for(const Option_p& opt : dynamic_cast<T*>(parent_)->options_)
+            if(opt.get() != this && opt->check_name(opt_name))
+                return excludes(opt.get());
+        throw IncorrectConstruction("Option " + opt_name + " is not defined");
+
+    }
+    /// Any number supported, any mix of string and Opt
+    template<typename A, typename B, typename... ARG>
+    Option* excludes(A opt, B opt1, ARG... args) {
         excludes(opt);
         return excludes(opt1, args...);
     }