+++ tests/AppTest.cpp
@@ -458,12 +458,69 @@ TEST_F(TApp, ExcludesFlags) {
     EXPECT_THROW(run(), CLI::ExcludesError);
 }
 
+TEST_F(TApp, ExcludesMixedFlags) {
+    CLI::Option* opt1 = app.add_flag("--opt1");
+    app.add_flag("--opt2");
+    CLI::Option* opt3 = app.add_flag("--opt3");
+    app.add_flag("--no")->excludes(opt1, "--opt2", opt3);
+
+    EXPECT_NO_THROW(run());
+
+    app.reset();
+    args = {"--no"};
+    EXPECT_NO_THROW(run());
+
+    app.reset();
+    args = {"--opt2"};
+    EXPECT_NO_THROW(run());
+
+    app.reset();
+    args = {"--no", "--opt1"};
+    EXPECT_THROW(run(), CLI::ExcludesError);
+
+    app.reset();
+    args = {"--no", "--opt2"};
+    EXPECT_THROW(run(), CLI::ExcludesError);
+}
 
 TEST_F(TApp, RequiresMultiFlags) {
     CLI::Option* opt1 = app.add_flag("--opt1");
     CLI::Option* opt2 = app.add_flag("--opt2");
     CLI::Option* opt3 = app.add_flag("--opt3");
     app.add_flag("--optall")->requires(opt1, opt2, opt3);
+
+    EXPECT_NO_THROW(run());
+
+    app.reset();
+    args = {"--opt1"};
+    EXPECT_NO_THROW(run());
+
+    app.reset();
+    args = {"--opt2"};
+    EXPECT_NO_THROW(run());
+
+    app.reset();
+    args = {"--optall"};
+    EXPECT_THROW(run(), CLI::RequiresError);
+
+    app.reset();
+    args = {"--optall", "--opt1"};
+    EXPECT_THROW(run(), CLI::RequiresError);
+
+    app.reset();
+    args = {"--optall", "--opt2", "--opt1"};
+    EXPECT_THROW(run(), CLI::RequiresError);
+
+    app.reset();
+    args = {"--optall", "--opt1", "--opt2", "--opt3"};
+    EXPECT_NO_THROW(run());
+}
+
+TEST_F(TApp, RequiresMixedFlags) {
+    CLI::Option* opt1 = app.add_flag("--opt1");
+    CLI::Option* opt2 = app.add_flag("--opt2");
+    CLI::Option* opt3 = app.add_flag("--opt3");
+    app.add_flag("--optall")->requires(opt1, "--opt2", "--opt3");
 
     EXPECT_NO_THROW(run());